refactor: 패키지명, 변수명 변경
