+++ src/main/java/kr/hhplus/be/server/application/order/OrderCriteria.java
@@ -1,7 +1,7 @@
 package kr.hhplus.be.server.application.order;
 
 import java.util.List;
-import kr.hhplus.be.server.web.order.request.OrderRequest;
+import kr.hhplus.be.server.interfaces.order.request.OrderRequest;
 import lombok.Builder;
 import lombok.Getter;
