--- src/main/java/kr/hhplus/be/server/config/JpaConfig.java
@@ -1,19 +1,0 @@
-package kr.hhplus.be.server.config;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
-import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
-import org.springframework.orm.jpa.JpaTransactionManager;
-import org.springframework.transaction.PlatformTransactionManager;
-
-@Configuration
-@EnableJpaAuditing
-@EnableJpaRepositories
-public class JpaConfig {
-
-    @Bean
-    public PlatformTransactionManager transactionManager() {
-        return new JpaTransactionManager();
-    }
-}
