+++ src/main/java/kr/hhplus/be/server/interfaces/coupon/request/CouponRequest.java
@@ -1,0 +1,15 @@
+package kr.hhplus.be.server.interfaces.coupon.request;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@AllArgsConstructor
+@Getter
+public class CouponRequest {
+
+    long userId;
+
+    long couponId;
+
+
+}
