+++ src/main/java/kr/hhplus/be/server/interfaces/order/response/OrderResponse.java
@@ -1,0 +1,20 @@
+package kr.hhplus.be.server.interfaces.order.response;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@Getter
+@AllArgsConstructor
+public class OrderResponse {
+
+    private long userId;
+
+    private long totalAmount;
+
+    private String orderAt;
+
+    public static OrderResponse of(long userId, long totalAmount,String orderAt) {
+        return new OrderResponse(userId, totalAmount, orderAt);
+    }
+
+}
