+++ src/main/java/kr/hhplus/be/server/interfaces/user/UserController.java
@@ -1,0 +1,29 @@
+package kr.hhplus.be.server.interfaces.user;
+
+import io.swagger.v3.oas.annotations.Parameter;
+import io.swagger.v3.oas.annotations.tags.Tag;
+import kr.hhplus.be.server.domain.user.User;
+import kr.hhplus.be.server.domain.user.UserService;
+import kr.hhplus.be.server.interfaces.ApiResult;
+import lombok.RequiredArgsConstructor;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequestMapping("/user")
+@Tag(name = "사용자 API")
+@RequiredArgsConstructor
+public class UserController {
+
+    private final UserService userService;
+
+    @GetMapping("/point/{userId}")
+    public ApiResult<UserResponse> getUserPoint(@Parameter(hidden = true) @PathVariable long userId) {
+        User user = userService.getUserInfo(userId);
+        return ApiResult.success(UserResponse.of(user.getUserId(), user.getPoint()));
+    }
+
+
+}
