--- src/main/java/kr/hhplus/be/server/web/coupon/request/CouponRequest.java
@@ -1,16 +1,0 @@
-package kr.hhplus.be.server.web.coupon.request;
-
-import kr.hhplus.be.server.domain.coupon.CouponCommand;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@AllArgsConstructor
-@Getter
-public class CouponRequest {
-
-    long userId;
-
-    long couponId;
-
-
-}
