--- src/main/java/kr/hhplus/be/server/web/product/response/ProductResponse.java
@@ -1,41 +1,0 @@
-package kr.hhplus.be.server.web.product.response;
-
-import com.fasterxml.jackson.annotation.JsonInclude;
-import java.util.List;
-import kr.hhplus.be.server.domain.product.ProductDTO.ProductOptionResult;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@Getter
-@AllArgsConstructor
-public class ProductResponse {
-
-    private long productId;
-    private String name;
-    private long price;
-    private String status;
-
-    @JsonInclude(JsonInclude.Include.NON_NULL)
-    private List<ProductOptionResponse> options;
-
-    public static ProductResponse of(long productId, String name, long price, String status, List<ProductOptionResponse> optionList) {
-        return new ProductResponse(productId, name, price, status, optionList);
-    }
-
-
-    public static ProductResponse from(ProductOptionResult info) {
-        List<ProductOptionResponse> optionList = info.getOptions().stream()
-            .map(ProductOptionResponse::from)
-            .toList();
-
-        return ProductResponse.of(
-            info.getProductId(),
-            info.getName(),
-            info.getPrice(),
-            info.getStatus().name(),
-            optionList
-        );
-    }
-
-
-}
